Switch to wireless34 for daisy
We're planning to use the 3.4 wireless stack on snow on kernel 3.8.
Switch to it right away so we get the most testing.

Even though daisy is still building kernel-3.4 by default it still
should be fine to include this flag since it will essentially be a
no-op.  I believe it will cause CONFIG_ATH9K_BTCOEX to be compiled as
a module on snow, but that shouldn't hurt.

BUG=chrome-os-partner:20117
TEST=Boot and download a big file over WiFi.
TEST=cmp vmlinux.uimg on 3.4 before and after this change and see no
important diffs.  Also check mwifiex modules and see no diffs.
TEST=See annoying "inform bss" message in serial console from 3.4 WiFi
stack:

Change-Id: I90ac2b21c9fac4bddbb17a0148f6ec56bddd189d
Reviewed-on: https://gerrit.chromium.org/gerrit/58332
Tested-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Doug Anderson <dianders@chromium.org>