puppy: Add dalmore subprofile
The puppy overlay emits builds for running on venice or dalmore.  By
default, the puppy overlay will emit builds for venice.  We define a
dalmore subprofile so that we can emit builds for dalmore.  Right now,
the dalmore subprofile amounts to selecting the "dalmore" use flag,
and replacing the "tegra114-venice_dtb" use flag (used by
chromiumos-overlay/sys-boot/chromeos-bootimage) with "tegra114-dalmore_dtb".
For everything else, the dalmore subprofile inherits everything from the
base profile.

BUG=none
TEST=Was able to emerge-puppy u-boot-next and chromeos-bootimage, for both
the base and the "dalmore" subprofile.  Tested that the dalmore subprofile
emits a build that works for dalmore.

Change-Id: I1af702b8e0813368cd53047665011c7428ee359a
Signed-off-by: Andrew Chew <achew@nvidia.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/58573
Reviewed-by: Mike Frysinger <vapier@chromium.org>