chromeos-bsp-daisy : Remove Atmel mXT Touchpad from wakeup_input_device_names
Due to crosbug.com/p/19227, enabling the Atmel mXT Touchpad for wakeup
may result in the system resuming as soon as we suspend when
the touchpad sends a spurious interrupt to the host. Lets disable this
until we get a fixed firmware from Atmel.

Signed-off-by: Benson Leung <bleung@chromium.org>

BUG=chrome-os-partner:20879
TEST=suspend a system with powerd_suspend from console.
Attempt to wake the system by touching touchpad.
Should not wake from touch.

Change-Id: I74bb0c80b911577cd2f79f45a294af3f857800dd
Reviewed-on: https://gerrit.chromium.org/gerrit/61787
Reviewed-by: Yufeng Shen <miletus@chromium.org>
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>
Commit-Queue: Benson Leung <bleung@chromium.org>
Tested-by: Benson Leung <bleung@chromium.org>