nyan: Add new nyan board overlay
This overlay will be used to build for venice2 boards.

BUG=none
TEST=./setup_board --board=nyan, and ./build_packages --board=puppy, and
saw that all package dependencies are well-formed.  Note that the kernel,
u-boot, and bct ebuilds currently fail.  The kernel and u-boot ebuilds fail
because chromeos-kernel-next and chromeos-u-boot don't yet have
tegra124/venice2 configs.  The bct ebuild fails because cbootimage needs
to be updated for t124 support.

Change-Id: I5a52d82d743aa3888196a6f2f5f63d2e984022d5
Signed-off-by: Andrew Chew <achew@nvidia.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/63769
Reviewed-by: Mike Frysinger <vapier@chromium.org>