Add a coreboot ebuild.
This ebuild satisfies the chromeos-coreboot virtual.

Also, instead of depending on the BCT package to install the BCT in the
firmware directory, my intention is to check the BCT config files into
coreboot and build the binary BCT as part of its build process. That will
make it easier for upstream to build a working coreboot image without chromeos
specific tools.

To support that, this ebuild depends on the cbootimage ebuild which installs
the cbootimage tool (provided by nvidia) for building BCTs instead of
depending on the package that would install the BCT source files.

BUG=None
TEST=With this and other changes, built coreboot based firmware for nyan.
BRANCH=None

Change-Id: I91d16034340cce78850d1bb19b0e29473daa529d
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/170855
Reviewed-by: Ronald Minnich <rminnich@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>