beaglebone: include packages for basic debug/use
This enables the basic ability to log in to a beaglebone Chrome OS
image, and to run commands like 'bash' and 'sudo'.  This also
includes any other packages that are known or suspected to be needed
for debugging problems in the lab.

CQ-DEPEND=CL:172304
BUG=chromium:304426
TEST=boot a beaglebone image, log in, run 'bash' and 'sudo'

Change-Id: Iad6016ef520ae129908911d40b05994cd9307b6e
Reviewed-on: https://chromium-review.googlesource.com/172114
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>