beaglebone:  Use chromeos-base/u-boot-scripts.
To support autoupdate, Chromium OS on arm boards provides two U-Boot
script files; one for slot A and one for slot B.  The scripts
provide the settings needed for U-Boot to select which slot to boot.
Postinst on non-Chrome arm hardware selects the slot to be booted
next by copying the appropriate script to the boot partition.

This updates the U-Boot default environment settings to load the
boot script and use its settings.

BUG=chromium:221802
TEST=boot an image

Change-Id: Ia6977dad11776bd9c16df2099bd03639aa5bbd02
Reviewed-on: https://chromium-review.googlesource.com/173180
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Queue: David James <davidjames@chromium.org>