thermal.sh: don't query nonexistent sensors
In some revisions of Snow we have no thermistors and thus no hwmon
devices.  Don't consider this to be an error.  If we need to confirm
that hwmon is working properly on devices that should have them, we
should add an autotest for that.

BUG=chrome-os-partner:22528
TEST=Manual:
  1. stop thermal                # to stop any existing script
  2. tail -f /var/log/messages&  # to show errors
  3. bash thermal.sh -d          # run manually with debugging
On rev4 hardware see printouts like:
  thermal.sh: Max CPU Freq set to 1700000 (Celsius: 34 30 32 31 32)
  08:54:08 , 1700000, 1700000 , 34, 31 31 30 31
On rev5 hardware see printouts like:
  thermal.sh: Max CPU Freq set to 1700000 (Celsius: 35 )
  08:53:13 , 1700000, 1700000 , 33,

Change-Id: Ie0562c6c2d52081c47ebb9f30df470c4b154222f
Reviewed-on: https://chromium-review.googlesource.com/174934
Reviewed-by: Katie Roberts-Hoffman <katierh@chromium.org>
Reviewed-by: Doug Anderson <dianders@chromium.org>
Commit-Queue: Doug Anderson <dianders@chromium.org>
Tested-by: Doug Anderson <dianders@chromium.org>