misc board profile cleanups
This fixes the repo_name for these boards:
	amd64-corei7
	puppy
	x86-mario

It also updates the peppy parent files to use the newer repo syntax.

We clean out make.defaults from the kernel-next subdir for a few boards
as we get the file from the base profile.

We can drop the smatch logic from existing package.use files because it
is enabled globally now for all boards.

Finally, we move the existing x86-mario package.use from the private
overlay to the public.

BUG=None
TEST=`cbuildbot {peppy,puppy,stumpy,x86-alex,x86-mario}-release` works

Change-Id: Idb44ce98e018fae89b695c22558e977f0c95db06
Reviewed-on: https://chromium-review.googlesource.com/177252
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>