Upgrade to new Cypress FW 2.15 for Daisy
We have a new FW drop from Cypress for Daisy.  This is version 2.15 and
the binary has been uploaded to the binary component server.  This CL
edits the ebuild to have the new values so it downloads the most updated
firmware when building for Daisy.

BUG=chrome-os-partner:24500
TEST=manually built for a Daisy with Cypress.  The new FW was picked up
and downloaded and installed in the right directory with the symlinks
set up from /lib/firmware.

Change-Id: I48ad5287dde85122ba8c82fc1a634dcb258c9854
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/179700
Reviewed-by: Benson Leung <bleung@chromium.org>