snow: Rename the chromeos-coreboot-snow ebuild to ...-daisy.
This makes the name of the ebuild consistent with the name of the overlay.
Because the ebuild has never existed with this name, I'm deleting the non-9999
version and letting the build infrastructure recreate it with the new name. It
should be ok to recycle revision numbers since they'll be treated
independently from the old set.

BUG=None
TEST=Built for daisy.
BRANCH=None

Change-Id: Iacfd4fb4c33fa4015f51b5d3e51c1a26bbf21a35
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/183567
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>