arm64-generic: Add simple overlay
This is mostly just a placeholder today, the overlay won't be able to
build packages and there is much, much work left to do to actually have
it runnable. We're going to start out with 32-bit userspace and a 64-bit
kernel in other overlays, But having a basic board enabled that we can
start working on top of for full 64-bit is still useful.

BUG=chromium:334368
TEST=none

Change-Id: I0f6dafeab0913b10c557d8175fa88674dea1aebd
Reviewed-on: https://chromium-review.googlesource.com/184041
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Olof Johansson <olofj@chromium.org>
Commit-Queue: Olof Johansson <olofj@chromium.org>