coreboot: Stop copying the BL1 in and use the one that's already there.
The 3rdparty directory in coreboot now has a repository under it, and that
repository now has a real BL1 in it which we can use for snow. We can now stop
copying the other one in and use the one that's already there.

BUG=None
TEST=Built for daisy/snow.
BRANCH=None

Change-Id: I33820f9ce1b5bd24c5ac5ccac0acea0fc74454bc
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/185615
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>