Add wakeup sources for Skate
Skate did non have a wakeup sources list like Spring and Daisy, which
made it impossible to wake with the keyboard/touchpad/etc.  This CL adds
the list of devices that can wake it up.

BUG=chrome-os-partner:25146
TEST=manually tested by first suspending with powerd_dbus_suspend,
observing that it could not be woken up by the touchpad.  I then merged
in this change, and repeated the process.  The touchpad could wake it up
after that.

Change-Id: I9ff5bad17d8b80c8ff21729b250c8687216f500a
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/186659
Reviewed-by: Benson Leung <bleung@chromium.org>