Populate rootdev device.
Removing guesswork in the code to find the destination device
for install/recovery/factory operation.

Need to populate the sysfs path [with wildcards if necessary]
for finding the fixed drive.

For intel chipset, use:
/sys/devices/pci0000:00/0000:00:1f.2/ata1/host*/target*/*/block/sd*
[* are necessary if the usb stick is found before the fixed device]

For arm based: based on the device tree of the machine:
/sys/devices/12200000.dwmmc0/mmc_host/mmc0/mmc0:0001/block/mmcblk0

BUG=chromium:343182,chromium:340761,chromium:336654
TEST=Install on link, falco and peach
TEST= - recovery
TEST= - factory wipe/shell
TEST= - chromeinstall from USB key
TEST= - cros flash <machine>

Change-Id: I32e6cf9244e7dc0f5fb1345652ca311ae0987790
Reviewed-on: https://chromium-review.googlesource.com/187427
Tested-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>