nyan: Fix udev rules to be less permissive
The current udev rules for nvrm seem to be a union of various iterations
of L4T.  Reduced the clutter to what's actually there right now.  In
addition, only allow rw by user and group (the default), and set the group
to "video", which has chronos and xorg as members.

BUG=chrome-os-partner:25440
TEST=Built ChromeOS for nyan with new rules file, and boot to ChromeOS
desktop.

Change-Id: Id229270f56df7ba2360e1e919890bb35c862eba1
Signed-off-by: Andrew Chew <achew@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/188128
Reviewed-by: Mike Frysinger <vapier@chromium.org>