Update Daisy VE1 touchpad config with extra noise immunity
This config was designed by Atmel Korea to address an electrical noise
issue that was being reported in Brazil.  As a side effect it seemed to
significantly mitigate another existing issue with 336s's touchpad.

There was an issue on 336s where the cursor would sometimes jump when you
had one finger already on the pad and introduced a second.  The two
fingers would merge briefly before splitting, which could result in a
jumping cursor.  This config mitigates the problem, making it harder to
trigger.

BUG=chrome-os-partner:26299
TEST=manually tested the config on 336s

Change-Id: I0baf5bba3a8da9c8047173c83e21a3b42ed29eb5
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/192121
Reviewed-by: Benson Leung <bleung@chromium.org>