drop -mtune=generic
When we have a -march value that is more specific than "generic", it makes
no sense to then use -mtune=generic.  We've already opted in to using insns
that are specific to a cpu, so telling gcc to optimize its scheduling for a
baseline cpu only serves to slow things down.

While we're here, also drop -mtune values that are redundant.  If you say
-march=foo, then -mtune defaults to foo.

BUG=chromium:367965
TEST=`cbuildbot {butterfly,falco,glimmer,leon,link,lumpy,panther,parrot,peppy,quawks,stout,stumpy,wolf,zako}-full` works

Change-Id: I82b2af72a20b8a6c9deda0c34f37e4a746d4a0ed
Reviewed-on: https://chromium-review.googlesource.com/197360
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Luis Lozano <llozano@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>