Glimmer: Update touch configuration
This changes:

0x08
Byte[6]~Byte[9] are anti-touch calibration setting and designed to load
two sets of different parameters dynamically. The setting “FF 01 00 00”
will be replaced with initial setting “05 32 19 C4” when timer is expired
(56th sec since start-up)

0x2F
This object is designed for passive stylus, currently it is disabled.

0x64
The setting 0x83 is to disable RPTENABLE.
The setting 0x3C is to enable DISSCRMSG0.

BUG=chrome-os-partner:28780
TEST=`emerge-glimmer chromeos-touch-config-glimmer`

Change-Id: Ib8d29d349244ba93ebe1946a428e13f59d3039c2
Reviewed-on: https://chromium-review.googlesource.com/199197
Reviewed-by: Yufeng Shen <miletus@chromium.org>
Commit-Queue: Yufeng Shen <miletus@chromium.org>
Tested-by: Yufeng Shen <miletus@chromium.org>