Moblab: Give sudo access to servod/reboot & Set-up Apache home dir
Not having to switch to user root to kick off servod and reboot will
be nice to have.

Also properly configures Apache's home directory to be able to ssh
into DUT's. This is required as the web frontend is now capable of
adding DUT's and detecting the appropriate labels.

BUG=chromium:351994
TEST=Able to launch servod, reboot moblab and control upstart
services.

Change-Id: I434f9daf4df27f3d6dc2b0e37813339afa7db931
Reviewed-on: https://chromium-review.googlesource.com/201667
Reviewed-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>