Clapper: Disable touchpad as a wake source in all cases
The touchpad should not be used to wake the device when in stand
mode. The only way to do this on Clapper is to diable the
touchpad as a wake source in all cases.

BUG=chrome-os-partner:28950
BRANCH=M36
TEST=Build and boot on Clapper. Log in and suspend machine using
'powerd_dbus_suspend'. Verify that touchpad does not wake device.

Change-Id: I07aff8e4543a58fda448837b8a7988cca3a1e3a5
Reviewed-on: https://chromium-review.googlesource.com/203260
Commit-Queue: Dave Parker <dparker@chromium.org>
Tested-by: Dave Parker <dparker@chromium.org>
Reviewed-by: Benson Leung <bleung@chromium.org>