veyron: Delete custom kernel ebuild (switches to 3.14)
This clobbers the virtual kernel build and switches us over to
building in the 3.14 tree.  If you still want to build the old kernel
you can do:

 cros_workon --board=veyron start chromeos-kernel-3_14
 cd ~/trunk/src/third_party/kernel/3.14/
 git remote add rockchip https://github.com/rkchrome/kernel.git
 git fetch rockchip
 git checkout rockchip/master

...though you'd need to re-checkout every time you do a repo sync.

NOTE: Unless we hold off on merging this for a while it's probably not
worth adding a blocker given the small number of people who have the
old kernel installed.  For those that run into an error they can
manually run:

 emerge-veyron --unmerge sys-kernel/rkchrome-kernel

BUG=chrome-os-partner:30167
TEST=See 3.14 kernel tree getting compiled

Change-Id: Ib0fbb375ef44f8963e3530345bc74f8e24cdeb53
Reviewed-on: https://chromium-review.googlesource.com/206402
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Commit-Queue: Doug Anderson <dianders@chromium.org>
Tested-by: Doug Anderson <dianders@chromium.org>