veyron: Choose kernel 3.14
We don't actually use the normal chromeos-kernel build since we
override with a custom kernel.  However, any work to get rockchip
working in our tree ought to happen in the 3.14 tree so let's pick
that one in preparation for deleting the custom kernel override here.

Checking this in now just to avoid conflicts with make.conf.

BUG=chrome-os-partner:30167
TEST=Delete custom kernel override and see 3.14 picked.

Change-Id: I82b900dc16ae82479e1ecaba28ab93e4222b4b47
Reviewed-on: https://chromium-review.googlesource.com/206401
Reviewed-by: Katie Roberts-Hoffman <katierh@chromium.org>
Commit-Queue: Doug Anderson <dianders@chromium.org>
Tested-by: Doug Anderson <dianders@chromium.org>