nyan_freon: Uprev chromeos-firmware to 3.5 so it is chosen over nyan's 3.0 version
When I submitted the last CL, I blindly followed the pre-upload hook's advice of
setting the version to 2.5. This caused my build to fail (I probably should have
tested it again). The reason the build failed was because our nyan_freon
chromeos-firmware-2.5 was a lower version than overlay-nyan-private's version 3.0.

The pre-upload script recommended 2.5 because nyan_freon is a public overlay variant.
So according to the table at [1], it should be versioned 2.5. The script, however,
does not take into account that the parent of nyan_freon is a private overlay, and
as such, nyan_freon effectively acts as a private overlay variant. In light of this,
We should really use 3.5 as our version number.

Also, while we're ignoring the pre-upload script, let's nuke the -r1 symlink b/c it's
not doing anything for us.

[1] - http://dev.chromium.org/chromium-os/how-tos-and-troubleshooting/portage-build-faq#TOC-Virtuals-and-central-management

BUG=None
TEST=nyan_freon build_packages works

Change-Id: If31503c18d45f2419880ae08fc365839a305d4a8
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/208911
Reviewed-by: Alex Sakhartchouk <alexst@chromium.org>