Moblab: Move off of using cros_disks to running mount.
cros-disks is dropping support for ext4 which Moblab uses to mount
the devserver's external storage.

This CL moves off of using cros-disks to mount the drive to using
the mount command instead.

BUG=chromium:315401
TEST=Verified that the external storage is mounted properly and the
devserver is running as expected. Ensured not having the drive
inserted fails as expected as well.

Change-Id: I09dda238e3babe82a62a4812fe0327ea55c211ce
Reviewed-on: https://chromium-review.googlesource.com/209153
Reviewed-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>