[moblab] boto file permissions & apache user reboot rights.
This change ensures that the boto file is owned by user moblab and
the moblab group has read/write access and creates an empty placeholder
file should the boto key not exist. This will be used by apache to
write in the boto key should one be uploaded via the AFE.

User apache now has sudo permissions to run reboot which is required
after updating autotest's config settings.

BUG=chromium:396694
TEST=After powerwash, a boto key can be uploaded. And verified that
if a boto key is placed manually, and the moblab is rebooted, the
same key exists with the appropriate permissions.

Change-Id: I7cdab17b9b28f3a3a8989c6c1a2e7a99c4044085
Reviewed-on: https://chromium-review.googlesource.com/212322
Reviewed-by: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: Simran Basi <sbasi@chromium.org>