Specify OZONE_PLATFORM_DEFAULT for freon boards
If gbm is working, the default is gbm. Otherwise, the default is dri.
Eventually gbm will be default everywhere.

This has no effect if the comand line flag is specified, it merely makes
manual invocation more ergonomic because the correct value is now the
default.

BUG=chromium:408434
CQ-DEPEND=CL:214680
TEST=build for link_freon, nyan_freon, daisy_freon

Change-Id: I71329ce61424ebbdcc2c01341d5615f774793fd4
Reviewed-on: https://chromium-review.googlesource.com/214691
Reviewed-by: Alex Sakhartchouk <alexst@chromium.org>
Tested-by: Michael Spang <spang@chromium.org>
Commit-Queue: Michael Spang <spang@chromium.org>