add arm-none-eabi toolchain for boards w/our EC
Boards that build the CrOS EC need this toolchain.  Previously they were
relying on it existing in the SDK by way of link, but as we change the
sdk to not install all toolchains by default, these boards need to pull
in the ones they use explicitly.

BUG=chromium:342199
TEST=pre-cq passes
TEST=building in new sdk works for these boards

Change-Id: I865a2243b293d06707d3d535c05ca512601aa9d5
Reviewed-on: https://chromium-review.googlesource.com/229821
Reviewed-by: Bill Richardson <wfrichar@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>