[mobbuild] Initialize buildtools checkout.
Adds an init script to checkout the buildtools layout on boot. This
layout provides the basic requirements for launching cbuildbot.

BUG=chromium:432620
TEST=buildtools is properly pulled down at boot. Also booted the system
without the network cord attached and observed it waiting for the
system to be online and once the cord was attached it successfully
pulled down the checkout.

Change-Id: Icb7ec7aaa7b4723972d5dd9b56e9288bcdb2a619
Reviewed-on: https://chromium-review.googlesource.com/231122
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>
Commit-Queue: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>