Candy: add arm-none-eabi toolchain
This is just another board added like:
https://chromium-review.googlesource.com/#/c/230001/

Boards that build the CrOS EC need this toolchain.  Previously they were
relying on it existing in the SDK by way of link, but as we change the
sdk to not install all toolchains by default, these boards need to pull
in the ones they use explicitly.

BUG=chromium:342199
TEST=None

Change-Id: I520580b7a452126334be1876ad278dda19f05e8e
Reviewed-on: https://chromium-review.googlesource.com/231702
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Bernie Thompson <bhthompson@chromium.org>
Tested-by: Bernie Thompson <bhthompson@chromium.org>