Expresso: add arm-none-eabi toolchain
This is just another board added like:
https://chromium-review.googlesource.com/#/c/230001/

Boards that build the CrOS EC need this toolchain. Previously they were
relying on it existing in the SDK by way of link, but as we change the
sdk to not install all toolchains by default, these boards need to pull
in the ones they use explicitly.

BUG=chromium:342199
TEST=image created and loaded to the platforms successfully

Change-Id: I0b5f8fa0b0c833180e7568cc65edff3473e19928
Reviewed-on: https://chromium-review.googlesource.com/232740
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Tedd An <tedd.an@intel.com>
Commit-Queue: Tedd An <tedd.an@intel.com>
Tested-by: Tedd An <tedd.an@intel.com>