Moblab: Add network bridge and restructure network configuration.
Adds bridge-utils & brctl to the Moblab Image.

Also restructured moblab-network-init to first configure the
network bridge and dhcpd and then bring up the usb-ethernet
adaptor and connect it to the bridge.

This allows containers to have network access and be on the same
test vlan as the DUTs, enabling them to run tests against the
test machines.

BUG=chromium:455877
TEST=Containers have network access and can access DUTs on the
test vlan.

Change-Id: Ifbcf9374c8c04551f6d99a179f2bd5183a5518d3
Reviewed-on: https://chromium-review.googlesource.com/248321
Reviewed-by: Thieu Le <thieule@chromium.org>
Reviewed-by: Jay Kim <yongjaek@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: Simran Basi <sbasi@chromium.org>