Add SAMA5D3 overlay
Based on the beaglebone overlay.

From the README:

This is a basic overlay for Atmel SAMA5D3 Xplained

http://www.atmel.com/tools/ATSAMA5D3-XPLD.aspx

This overlay will build an image suitable to boot from SD card on this
platform. Install to NAND is not supported at this time.

There is no UI/Chrome on this board.

BUG=none
TEST=setup_board, build packages and build image

Change-Id: I0bb0a6a14c0fc3e7e2225e14875a9a68ff8e031a
Reviewed-on: https://chromium-review.googlesource.com/250990
Trybot-Ready: Olof Johansson <olofj@chromium.org>
Tested-by: Olof Johansson <olofj@chromium.org>
Reviewed-by: Katie Roberts-Hoffman <katierh@chromium.org>
Commit-Queue: Olof Johansson <olofj@chromium.org>