Add cpufreq.conf for veyron
Use this config file to set the hispeed frequency.  This is the
frequency that the system will "boost" to when there is user
interaction.  It also serves as the minimum frequency to run at when the
system load approaches 1.0.

BUG=chrome-os-partner:37673
TEST=watch cpu frequency scale while idle, playing you tube, and
watching the web GL aquarium.

Change-Id: Ie2dd84b111e013f9562ef2961c0babf7a9956616
Reviewed-on: https://chromium-review.googlesource.com/257921
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Dylan Reid <dgreid@chromium.org>
Tested-by: Dylan Reid <dgreid@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>