gizmo: move to kernel 3.14
This bumps the kernel up to 3.14, which appears to be working. This
rebuilds a static kernel configuration, based on the 3.8 panther-minimal
static config with similiar areas removed from the config to reduce the
numbers of peripheral-support modules that would be built. Also tunes
fore core2 instead of i7, as gizmo isn't expected to be high-end.

BUG=brillo:481
TEST=`service ping` shows "ServiceManager is ready"
TEST=/dev/pstore/console-ramoops exists after soft reboot

Change-Id: Ie3edbf541e532b75f0fc7ca49736d003584bca04
Signed-off-by: Kees Cook <keescook@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/259835
Reviewed-by: Lee Campbell <leecam@chromium.org>