Set USE flags for mesa
These USE flags are not needed for a real oak build. It is to get the build
running so we can set up a builder before we check in all the graphics
libraries. The UI does not work until we have the correct graphics libraries.

BUG=chrome-os-partner:37408
TEST=Run `build_packages --board oak && build_image --board oak test`
     Verify the test image boots on my Oak.

Change-Id: I3d98bd179fe35e524b152536cd8c461bbb5c2328
Reviewed-on: https://chromium-review.googlesource.com/267401
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Ricky Liang <jcliang@chromium.org>
Trybot-Ready: Ricky Liang <jcliang@chromium.org>
Tested-by: Ricky Liang <jcliang@chromium.org>