Disable USE=X for chromeos-chrome in project-freon (reland)
We had this enabled to pull in X libs to to the Chrome SDK as a
transitional measure while we worked on removing the remaining
dependencies.

We finished that work a while ago, so it is no longer necessary to have
these in the SDK on freon builds. So turn this off during the next PFQ
run to save space and build time.

BUG=chromium:484324
TEST=build_packages --board link chromeos-chrome [X enabled]
     setup_board --board link --force
     cros_workon --board link start chromeos-chrome
     build_packages --board link chromeos-chrome [X disabled]

Change-Id: I6de9cbf7b9f3f4cc0fbbe29261b0420d07fc2b16
Reviewed-on: https://chromium-review.googlesource.com/273264
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Alex Sakhartchouk <alexst@chromium.org>
Tested-by: Michael Spang <spang@chromium.org>