Disable USE=X for chromeos-chrome in project-freon
We had this enabled to pull in X libs to to the Chrome SDK as a
transitional measure while we worked on removing the remaining
dependencies.

We finished that work a while ago, so it is no longer necessary to have
these in the SDK on freon builds. So turn this off during the next PFQ
run to save space and build time.

BUG=chromium:484324
TEST=build_packages --board link chromeos-chrome [X enabled]
     setup_board --board link --force
     cros_workon --board link start chromeos-chrome
     build_packages --board link chromeos-chrome [X disabled]

Change-Id: Ic3731573ab554c5a72285f9c97fb3f72243bb76e
Reviewed-on: https://chromium-review.googlesource.com/269135
Reviewed-by: Michael Spang <spang@chromium.org>
Commit-Queue: Michael Spang <spang@chromium.org>
Tested-by: Michael Spang <spang@chromium.org>