Fixup chromeos-chrome conditional
The PFQ landed a new chrome while this flag change was in the commit
queue.

This chumps in the right version to get prebuilts for chrome back for the
last PFQ cycle, while ensuring the next ones picks up this flag flip. It
seems that the only way to make these changes is to chump them at just
the right moment.

BUG=chromium:484324
TEST=build_packages --board link chromeos-chrome [X enabled]
     setup_board --board link --force
     cros_workon --board link start chromeos-chrome
     build_packages --board link chromeos-chrome [X disabled]

Change-Id: I343cad36349e15be3a995843d42864b51dd180b3
Reviewed-on: https://chromium-review.googlesource.com/271378
Reviewed-by: Michael Spang <spang@chromium.org>
Commit-Queue: Michael Spang <spang@chromium.org>
Tested-by: Michael Spang <spang@chromium.org>