lulu: turn on wake-on-wifi and dark resume
We want to activate the wake-on-wifi feature, so we need to build
images with this USE flag to compile the necessary functionality
into shill. This also adds dark resume preferences so that the system
does not come up in full resume whenever we wake up for lucid sleep.

BUG=chromium:476177
TEST=build and boot, check shill device properties for /device/wlan0,
  activate_short_dark_resume and check that the device dark resumes
  correctly

Change-Id: Ic00331ce01e6b3e529eac5f0290c0923898d5f2d
Reviewed-on: https://chromium-review.googlesource.com/271628
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>
Commit-Queue: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>