overlay-oak: Add img-ddk virtual, pulling in img-ddk-bin
This is required so that media-libs/mesa-img can pull in the
correct dependency (img-ddk or img-ddk-bin). We cannot use
"internal" in this context, as we do not want to set that
USE flag in our partner overlay chromeos-overlay-oak.

BUG=chromium:439783
TEST=Build oak image on public repository (with required kernel
     patches), oak-rev1 boots to UI.

Change-Id: I4fb6344825345c4da691d96dc9075a639fbbe809
Reviewed-on: https://chromium-review.googlesource.com/273080
Tested-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Nicolas Boichat <drinkcat@chromium.org>
Trybot-Ready: Nicolas Boichat <drinkcat@chromium.org>